Refactoring workers and serializers
