+++ src/BenchmarksClient/Workers/WorkerFactory.cs
@@ -5,24 +5,20 @@
 using System;
 using System.Collections.Generic;
 using Benchmarks.ClientJob;
-using BenchmarksWorkers.Workers;
 
-namespace BenchmarksWorkers
+namespace BenchmarksClient.Workers
 {
     static public class WorkerFactory
     {
         public static Dictionary<Worker, Func<ClientJob, IWorker>> Workers = new Dictionary<Worker, Func<ClientJob, IWorker>>();
-        public static Dictionary<Worker, Func<IResultsSerializer>> ResultSerializers = new Dictionary<Worker, Func<IResultsSerializer>>();
 
         static WorkerFactory()
         {
             // Wrk
             Workers[Worker.Wrk] = clientJob => new WrkWorker(clientJob);
-            ResultSerializers[Worker.Wrk] = () => new WrkSerializer();
 
             // SignalR
             Workers[Worker.SignalR] = clientJob => new SignalRWorker(clientJob);
-            ResultSerializers[Worker.SignalR] = () => new SignalRSerializer();
         }
 
         static public IWorker CreateWorker(ClientJob clientJob)
@@ -30,16 +26,6 @@ namespace BenchmarksWorkers
             if (Workers.TryGetValue(clientJob.Client, out var workerFactory))
             {
                 return workerFactory(clientJob);
-            }
-
-            return null;
-        }
-
-        static public IResultsSerializer CreateResultSerializer(ClientJob clientJob)
-        {
-            if (ResultSerializers.TryGetValue(clientJob.Client, out var serializerFactory))
-            {
-                return serializerFactory();
             }
 
             return null;
