+++ src/BenchmarksDriver/Serializers/SignalRSerializer.cs
@@ -10,7 +10,7 @@ using Benchmarks.ClientJob;
 using Benchmarks.ServerJob;
 using Newtonsoft.Json;
 
-namespace BenchmarksWorkers.Workers
+namespace BenchmarksDriver.Serializers
 {
     class SignalRSerializer : IResultsSerializer
     {
