--- src/BenchmarksWorkers/BenchmarksWorkers.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <LangVersion>7.1</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="$(JsonNetVersion)" />
-    <PackageReference Include="System.Net.Http" Version="4.3.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Benchmarks.ClientJob\Benchmarks.ClientJob.csproj" />
-    <ProjectReference Include="..\Benchmarks.ServerJob\Benchmarks.ServerJob.csproj" />
-    <PackageReference Include="System.Data.SqlClient" Version="$(CoreFxVersion)" />
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="$(SignalRVersion)" />
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client.MsgPack" Version="$(SignalRVersion)" />
-  </ItemGroup>
-</Project>
